+++ Light.hpp
@@ -7,7 +7,7 @@
 
 class Light {
 public:
-    glm::vec3 pos;
+    glm::dvec3 pos;
     glm::vec3 color;
     float intensity;
