+++ RenderingTask.hpp
@@ -66,7 +66,8 @@ private:
 
     Ray getPrimaryRay(unsigned int px, unsigned int py) const;
     glm::vec3 traceRay(const Ray &r, unsigned int maxDepth) const;
-    bool findNearestIntersection(const Ray &r, float &t, glm::vec3 &n, const Material **mat) const;
+    bool findNearestIntersection(const Ray &r, double &t, glm::dvec3 &n,
+                                 const Material **mat) const;
     bool isObstructed(const Ray &r, const Light &l) const;
     void renderBatch(std::vector<unsigned char> &imgData, unsigned int from, unsigned int count,
                      unsigned int &progress) const;
